+++ core_helpers/Base_Page.py
@@ -5,11 +5,11 @@ There are useful wrappers for common Selenium operations
 
 from selenium.webdriver.common.by import By
 import os,inspect
-from .drivers.driverfactory import DriverFactory
-from .core_helpers.selenium_action_objects import Selenium_Action_Objects
-from .core_helpers.remote_objects import Remote_Objects
-from .core_helpers.logging_objects import Logging_Objects
-from .core_helpers.screenshot_objects import Screenshot_Objects
+from core_helpers.drivers.driverfactory import DriverFactory
+from .selenium_action_objects import Selenium_Action_Objects
+from .remote_objects import Remote_Objects
+from .logging_objects import Logging_Objects
+from .screenshot_objects import Screenshot_Objects
 from page_objects import PageFactory
 import conf.base_url_conf
 from utils import accessibility_util
