+++ core_helpers/screenshot_objects.py
