+++ page_objects/bitcoin_main_page.py
@@ -6,7 +6,7 @@ import sys
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 import conf.locators_conf as locators
 from utils.Wrapit import Wrapit
-from .Mobile_Base_Page import Mobile_Base_Page
+from ..core_helpers.Mobile_Base_Page import Mobile_Base_Page
 
 
 class Bitcoin_Main_Page(Mobile_Base_Page):
