+++ page_objects/contact_page.py
@@ -3,7 +3,7 @@ This class models the Contact page.
 URL: contact
 The page consists of a header, footer and form object.
 """
-from .Base_Page import Base_Page
+from core_helpers.Base_Page import Base_Page
 from .contact_form_object import Contact_Form_Object
 from .header_object import Header_Object
 from .footer_object import Footer_Object
