+++ page_objects/tutorial_main_page.py
@@ -3,7 +3,7 @@ This class models the main Selenium tutorial page.
 URL: selenium-tutorial-main
 The page consists of a header, footer, form and table objects
 """
-from .Base_Page import Base_Page
+from core_helpers.Base_Page import Base_Page
 from .form_object import Form_Object
 from .header_object import Header_Object
 from .table_object import Table_Object
