+++ page_objects/tutorial_redirect_page.py
@@ -3,7 +3,7 @@ This class models the redirect page of the Selenium tutorial
 URL: selenium-tutorial-redirect
 The page consists of a header, footer and some text
 """
-from .Base_Page import Base_Page
+from core_helpers.Base_Page import Base_Page
 from .header_object import Header_Object
 from .footer_object import Footer_Object
 import conf.locators_conf as locators
