+++ page_objects/zero_mobile_page.py
@@ -3,7 +3,7 @@ This class models the first dummy page needed by the framework to start.
 URL: None
 Please do not modify or delete this page
 """
-from .Mobile_Base_Page import Mobile_Base_Page
+from ..core_helpers.Mobile_Base_Page import Mobile_Base_Page
 
 class Zero_Mobile_Page(Mobile_Base_Page):
     "Page Object for the dummy page"
