+++ page_objects/zero_page.py
@@ -3,7 +3,7 @@ This class models the first dummy page needed by the framework to start.
 URL: None
 Please do not modify or delete this page
 """
-from .Base_Page import Base_Page
+from core_helpers.Base_Page import Base_Page
 
 class Zero_Page(Base_Page):
     "Page Object for the dummy page"
